tests/requirements.txt
nose
nose-progressive
nose-testconfig
nose_parameterized==0.3.3
requests==2.11.1
gitpython
